Fixed Error in 01_Transfer_Incremental_Inventory_Unit
Fixed Error in 01_Transfer_Incremental_Inventory_Unit which caused some object types just to be filled with Dummy values.
This error occured when Pyspark could not determine the Type of a field. In case the type is unclear now it will be casted to String.
Error was found by Artur König. He also helped to debug it